Xe133m simulation for geant4
Simple simulation files for geant4 using gears. After running the run1.mac the steps are combined using the combineStepPointsTohits.C to emulate an accurate detector response. Finally the data is sorted using the sort_events.cpp. Here histograms are generated and coincidence logic is defined to allow understanding for detector responses for determining radiation interactions with the detectors.